# 后端代码/api/openapi.yaml
openapi: 3.0.3
info:
  title: 智慧党建系统 API
  description: |
    智慧党建系统后端API接口文档
    
    ## 认证方式
    本系统使用JWT（JSON Web Token）进行身份认证，支持两种登录方式：
    - **用户端**：微信小程序登录（通过OpenID）
    - **管理端**：用户名密码登录
    
    ## 请求头
    除登录接口外，所有需要认证的接口都需要在请求头中添加：
    ```
    Authorization: Bearer {JWT_TOKEN}
    ```
    
    ## 响应格式
    所有接口都返回统一的响应格式：
    ```json
    {
      "code": 200,
      "message": "操作成功",
      "data": {}
    }
    ```
    
  version: 1.0.0
  contact:
    name: 智慧党建系统开发团队
    email: dev@smart-party.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 开发环境
  - url: https://api.smart-party.com
    description: 生产环境
...
paths:
...
  /auth/user/login:
    post:
      tags:
        - 认证管理
      summary: 用户端登录
      description: 微信小程序用户通过OpenID登录
      operationId: userLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginDTO'
            example:
              openId: "wx_openid_123456"
              username: "张三"
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
                properties:
                  data:
                    $ref: '#/components/schemas/LoginVO'
              example:
                code: 200
                message: "登录成功"
                data:
                  userId: 1
                  username: "张三"
                  userType: "MEMBER"
                  accessToken: "eyJhbGciOiJIUzI1NiJ9..."
                  refreshToken: "eyJhbGciOiJIUzI1NiJ9..."
                  expiresIn: 7200
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 400
                message: "OpenID不能为空"
                data: null
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 401
                message: "用户不存在"
                data: null
  /auth/admin/login:
    post:
      tags:
        - 认证管理
      summary: 管理端登录
      description: 管理员通过用户名和密码登录
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginDTO'
            example:
              username: "admin"
              password: "123456"
              captcha: "1234"
              captchaId: "captcha_123"
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
                properties:
                  data:
                    $ref: '#/components/schemas/LoginVO'
              example:
                code: 200
                message: "登录成功"
                data:
                  userId: 1
                  username: "管理员"
                  userType: "ADMIN"
                  accessToken: "eyJhbGciOiJIUzI1NiJ9..."
                  refreshToken: "eyJhbGciOiJIUzI1NiJ9..."
                  expiresIn: 7200
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 400
                message: "用户名不能为空"
                data: null
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 401
                message: "用户名或密码错误"
                data: null
  /auth/refresh:
    post:
      tags:
        - 认证管理
      summary: 刷新访问令牌
      description: 使用刷新令牌获取新的访问令牌
      operationId: refreshToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDTO'
            example:
              refreshToken: "eyJhbGciOiJIUzI1NiJ9..."
      responses:
        '200':
          description: 令牌刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 200
                message: "令牌刷新成功"
                data: "eyJhbGciOiJIUzI1NiJ9..."
        '401':
          description: 刷新令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 401
                message: "刷新令牌已过期"
                data: null
  /auth/logout:
    post:
      tags:
        - 认证管理
      summary: 用户登出
      description: 用户登出，将当前令牌加入黑名单
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 200
                message: "登出成功"
                data: "登出成功"
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 401
                message: "缺少认证令牌"
                data: null
  /auth/profile:
    get:
      tags:
        - 认证管理
      summary: 获取当前用户信息
      description: 获取当前登录用户的详细信息
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
                properties:
                  data:
                    $ref: '#/components/schemas/UserProfile'
              example:
                code: 200
                message: "获取成功"
                data:
                  id: 1
                  name: "张三"
                  type: "MEMBER"
                  auth: "USER"
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
              example:
                code: 401
                message: "令牌已过期"
                data: null

  # ==================== 用户管理 ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        在请求头中添加JWT令牌：
        ```
        Authorization: Bearer {JWT_TOKEN}
        ```

# ==================== 数据模型 ====================
  schemas:
...
    Result:
      type: object
      properties:
        code:
          type: integer
          description: 响应状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: "操作成功"
        data:
          description: 响应数据
      required:
        - code
        - message

    # 分页结果模型
...
    UserLoginDTO:
      type: object
      properties:
        openId:
          type: string
          description: 微信OpenID
          example: "wx_openid_123456"
        username:
          type: string
          description: 用户名
          example: "张三"
      required:
        - openId
    AdminLoginDTO:
      type: object
      properties:
        username:
          type: string
          description: 用户名
          example: "admin"
        password:
          type: string
          description: 密码
          example: "123456"
        captcha:
          type: string
          description: 验证码
          example: "1234"
        captchaId:
          type: string
          description: 验证码ID
          example: "captcha_123"
      required:
        - username
        - password
    RefreshTokenDTO:
      type: object
      properties:
        refreshToken:
          type: string
          description: 刷新令牌
          example: "eyJhbGciOiJIUzI1NiJ9..."
      required:
        - refreshToken
    LoginVO:
      type: object
      properties:
        userId:
          type: integer
          format: int64
          description: 用户ID
          example: 1
        username:
          type: string
          description: 用户名
          example: "张三"
        userType:
          type: string
          description: 用户类型
          enum: [MEMBER, ADMIN]
          example: "MEMBER"
        accessToken:
          type: string
          description: 访问令牌
          example: "eyJhbGciOiJIUzI1NiJ9..."
        refreshToken:
          type: string
          description: 刷新令牌
          example: "eyJhbGciOiJIUzI1NiJ9..."
        expiresIn:
          type: integer
          description: 令牌过期时间（秒）
          example: 7200
    UserProfile:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 用户ID
          example: 1
        name:
          type: string
          description: 用户名
          example: "张三"
        type:
          type: string
          description: 用户类型
          example: "MEMBER"
        auth:
          type: string
          description: 认证类型
          example: "USER"

    # 用户相关模型
...